Tests refactoring #8
I have refactored my tests to reduce duplicate code and to make future test writing simpler and more productive. There are now some base classes for the various different kinds of tests that I need. CrmUnit class is for unit tests that mock out the dependency on CRM, CrmIntegration class is for tests that perform integration with CRM, and then there is a derived SyncIntegration tests class to serve as a base class for all tests that will perform sync integration testing.